fix: change uint32 type to int64 type

+++ tests/history/generate_csvs.py
@@ -29,11 +29,11 @@ def generate_daily_test_data(
     days = cal.days_in_range(first_day, last_day)
 
     days_count = len(days)
-    o = np.zeros(days_count, dtype=np.uint32)
-    h = np.zeros(days_count, dtype=np.uint32)
-    l = np.zeros(days_count, dtype=np.uint32)  # noqa: E741
-    c = np.zeros(days_count, dtype=np.uint32)
-    v = np.zeros(days_count, dtype=np.uint32)
+    o = np.zeros(days_count, dtype=np.int64)
+    h = np.zeros(days_count, dtype=np.int64)
+    l = np.zeros(days_count, dtype=np.int64)  # noqa: E741
+    c = np.zeros(days_count, dtype=np.int64)
+    v = np.zeros(days_count, dtype=np.int64)
 
     last_open = starting_open * 1000
     last_volume = starting_volume
@@ -97,11 +97,11 @@ def generate_minute_test_data(
     cal = get_calendar("XNYS")
     minutes = cal.sessions_minutes(first_day, last_day)
 
-    o = np.zeros(minutes_count, dtype=np.uint32)
-    h = np.zeros(minutes_count, dtype=np.uint32)
-    l = np.zeros(minutes_count, dtype=np.uint32)  # noqa: E741
-    c = np.zeros(minutes_count, dtype=np.uint32)
-    v = np.zeros(minutes_count, dtype=np.uint32)
+    o = np.zeros(minutes_count, dtype=np.int64)
+    h = np.zeros(minutes_count, dtype=np.int64)
+    l = np.zeros(minutes_count, dtype=np.int64)  # noqa: E741
+    c = np.zeros(minutes_count, dtype=np.int64)
+    v = np.zeros(minutes_count, dtype=np.int64)
 
     last_open = starting_open * 1000
     last_volume = starting_volume
